+++ src/rovr/variables/constants.py
@@ -1,5 +1,7 @@
 import platform
 from dataclasses import dataclass
+from os import environ
+from shutil import which
 from typing import Literal
 
 from textual.binding import Binding, BindingType
@@ -12,6 +14,26 @@ if "config" not in globals():
     global config, schema
     schema, config = load_config()
     config_setup()
+
+if "file" not in globals():
+    global file
+    # check for $ROVR_FILE_ONE
+    if (  # noqa: SIM114
+        "ROVR_FILE_ONE" in environ
+        and (found := which(environ["ROVR_FILE_ONE"])) is not None
+    ):
+        file = found
+    # check for $YAZI_FILE_ONE
+    elif (  # noqa: SIM114
+        "YAZI_FILE_ONE" in environ
+        and (found := which(environ["YAZI_FILE_ONE"])) is not None
+    ):
+        file = found
+    # check for `file` existence
+    elif (found := which("file")) is not None:
+        file = found
+    else:
+        file = None
 
 
 @dataclass